refactor: change simple components for custom components with based styles and properties and adjust imports
