+++ packages/components/ui/carousel.tsx
@@ -7,7 +7,7 @@ import useEmblaCarousel, {
 import * as React from "react";
 
 import { Button } from "@/packages/components/ui/button";
-import { cn } from "@/packages/utils/cn";
+import { cn } from "@/utils/cn";
 
 type CarouselApi = UseEmblaCarouselType[1];
 type UseCarouselParameters = Parameters<typeof useEmblaCarousel>;
@@ -206,7 +206,7 @@ const CarouselPrevious = React.forwardRef<
       variant={variant}
       size={size}
       className={cn(
-        "absolute  h-8 w-8 rounded-full",
+        "absolute h-8 w-8 rounded-full",
         orientation === "horizontal"
           ? "-left-12 top-1/2 -translate-y-1/2"
           : "-top-12 left-1/2 -translate-x-1/2 rotate-90",
