+++ packages/components/ui/tooltip.tsx
@@ -3,7 +3,7 @@
 import * as TooltipPrimitive from "@radix-ui/react-tooltip";
 import * as React from "react";
 
-import { cn } from "@/packages/utils/cn";
+import { cn } from "@/utils/cn";
 
 const TooltipProvider = TooltipPrimitive.Provider;
