+++ ui/journey/journey-connecting-line.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/packages/utils/cn";
+import { cn } from "@/utils/cn";
 import { HTMLAttributes } from "react";
 
 interface JourneyConnectingLineProps extends HTMLAttributes<HTMLDivElement> {}
@@ -15,7 +15,7 @@ export const JourneyConnectingLine = ({
           "pointer-events-none h-full w-px border-l-[1px] border-gray-200 dark:border-gray-700/50",
           className,
         )}
-      ></div>
+      />
     </div>
   );
 };
